Refactored handles to make stats independent of tracing.
