+++ stats-app-sample/init.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"log"
 	"time"
 
 	"github.com/google/instrumentation-go/stats"
@@ -10,9 +11,12 @@ import (
 )
 
 // Create/register view description of type gauge float
-var c = make(chan *stats.View, 256)
+var (
+	RPCserverServerElapsedTime     stats.MeasureDescFloat64
+	RPCserverServerElapsedTimeDist *stats.DistributionViewDesc
+)
 
-func registerCanonical() {
+func registerViews() {
 	// ---------------------------------------MEASURES---------------------------
 	// Create/register measure descriptions of type string
 	muString := &stats.MeasurementUnit{}
@@ -55,12 +59,11 @@ func registerCanonical() {
 	// ---------------------------------------VIEWS---------------------------
 	// Creates keys/views
 	mgr := tagging.DefaultKeyManager()
-
-	k1, err := mgr.CreateKeyStringUTF8("key1")
+	k1, err := mgr.CreateKeyInt64("key1")
 	if err != nil {
 		panic(fmt.Sprintf("Key k1 not created %v", err))
 	}
-	k2, err := mgr.CreateKeyStringUTF8("key2")
+	k2, err := mgr.CreateKeyInt64("key2")
 	if err != nil {
 		panic(fmt.Sprintf("Key k2 not created %v", err))
 	}
@@ -68,6 +71,9 @@ func registerCanonical() {
 	if err != nil {
 		panic(fmt.Sprintf("Key k3 not created %v", err))
 	}
+
+	// Create/register view description of type gauge float
+	c := make(chan *stats.View, 256)
 
 	vwGaugeS := &stats.GaugeStringViewDesc{
 		Vdc: &stats.ViewDescCommon{
@@ -187,9 +193,9 @@ func registerCanonical() {
 	}
 	// ---------------------------------------CREATES TAGS---------------------------
 	// Set tags values in mutations
-	mut1 := k1.CreateMutation("s1", tagging.BehaviorAdd)
-	mut2 := k2.CreateMutation("s2", tagging.BehaviorAdd)
-	mut3 := k3.CreateMutation("s3", tagging.BehaviorAdd)
+	mut1 := k1.CreateMutation(1, tagging.BehaviorAdd)
+	mut2 := k2.CreateMutation(1, tagging.BehaviorAdd)
+	mut3 := k3.CreateMutation("s1", tagging.BehaviorAdd)
 	// ---------------------------------------RECORD USAGE---------------------------
 	ctx := tagging.NewContextWithMutations(context.Background(), mut1, mut2, mut3)
 
@@ -199,17 +205,17 @@ func registerCanonical() {
 	// Creates few measurements
 	m1 := mDescString.CreateMeasurement("HIGH")
 	m2 := mDescBool.CreateMeasurement(true)
-	m3 := mDescFloat64.CreateMeasurement(1)
-	m4 := mDescFloat64.CreateMeasurement(1)
-	m5 := mDescInt64.CreateMeasurement(1)
-	m6 := mDescInt64.CreateMeasurement(1)
+	m3 := mDescFloat64.CreateMeasurement(209)
+	m4 := mDescFloat64.CreateMeasurement(9)
+	m5 := mDescInt64.CreateMeasurement(20)
+	m6 := mDescInt64.CreateMeasurement(40)
 	stats.RecordMeasurements(ctx, m1, m2, m3, m4, m5, m6)
 
 	// ---------------------------------------CREATES TAGS---------------------------
 	// Set tags values in mutations
-	mut1 = k1.CreateMutation("s1", tagging.BehaviorAdd)
-	mut2 = k2.CreateMutation("s2", tagging.BehaviorAdd)
-	mut3 = k3.CreateMutation("s3", tagging.BehaviorAdd)
+	mut1 = k1.CreateMutation(2, tagging.BehaviorAdd)
+	mut2 = k2.CreateMutation(2, tagging.BehaviorAdd)
+	mut3 = k3.CreateMutation("s2", tagging.BehaviorAdd)
 	// ---------------------------------------RECORD USAGE---------------------------
 	ctx = tagging.NewContextWithMutations(context.Background(), mut1, mut2, mut3)
 
@@ -218,16 +224,53 @@ func registerCanonical() {
 	//...
 	// Create context with mutations
 	// Creates few measurements
-	// m1 = mDescString.CreateMeasurement("LOW")
-	// m2 = mDescBool.CreateMeasurement(false)
-	// m3 = mDescFloat64.CreateMeasurement(1)
-	// m4 = mDescFloat64.CreateMeasurement(1)
-	// m5 = mDescInt64.CreateMeasurement(1)
-	// m6 = mDescInt64.CreateMeasurement(1)
-	go func() {
-		for {
-			stats.RecordMeasurements(ctx, m1, m2, m3, m4, m5, m6)
-			time.Sleep(time.Second * 5)
+	m1 = mDescString.CreateMeasurement("LOW")
+	m2 = mDescBool.CreateMeasurement(false)
+	m3 = mDescFloat64.CreateMeasurement(218)
+	m4 = mDescFloat64.CreateMeasurement(18)
+	m5 = mDescInt64.CreateMeasurement(5)
+	m6 = mDescInt64.CreateMeasurement(10)
+	stats.RecordMeasurements(ctx, m1, m2, m3, m4, m5, m6)
+}
+
+func registerRpcCanonical() {
+	var measures []stats.MeasureDesc
+	var views []stats.ViewDesc
+
+	rpcMillisBucketBoundaries := []float64{0, 1, 2, 3, 4, 5, 6, 8, 10, 13, 16, 20, 25, 30, 40, 50, 65, 80, 100, 130, 160, 200, 250, 300, 400, 500, 650, 800, 1000, 2000, 5000, 10000, 20000, 50000, 100000}
+
+	keyMethod, err := tagging.DefaultKeyManager().CreateKeyStringUTF8("grpc.method")
+	if err != nil {
+		log.Fatalf("init() failed to create/retrieve keyStringUTF8. %v", err)
+	}
+
+	milliseconds := &stats.MeasurementUnit{
+		Power10:    -3,
+		Numerators: []stats.BasicUnit{stats.SecsUnit},
+	}
+
+	RPCserverServerElapsedTime = stats.NewMeasureDescFloat64("/rpc/server/server_elapsed_time", "Server elapsed time in msecs", milliseconds)
+	measures = append(measures, RPCserverServerElapsedTime)
+	for _, m := range measures {
+		if err := stats.RegisterMeasureDesc(m); err != nil {
+			log.Fatalf("init() failed to register %v.\n %v", m, err)
 		}
-	}()
+	}
+
+	RPCserverServerElapsedTimeDist = &stats.DistributionViewDesc{
+		Vdc: &stats.ViewDescCommon{
+			Name:            "/rpc/server/server_elapsed_time/distribution_cumulative",
+			Description:     "Server elapsed time in msecs",
+			MeasureDescName: RPCserverServerElapsedTime.Meta().Name(),
+			TagKeys:         []tagging.Key{keyMethod},
+		},
+		Bounds: rpcMillisBucketBoundaries,
+	}
+	views = append(views, RPCserverServerElapsedTimeDist)
+
+	for _, v := range views {
+		if err := stats.RegisterViewDesc(v); err != nil {
+			log.Fatalf("init() failed to register %v.\n %v", v, err)
+		}
+	}
 }
